+++ pkg/ingester/ingester_ring_test.go
@@ -6,9 +6,11 @@ import (
 	"testing"
 	"time"
 
+	"github.com/go-kit/log"
 	"github.com/grafana/dskit/flagext"
 	"github.com/grafana/dskit/ring"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestRingConfig_DefaultConfigToLifecyclerConfig(t *testing.T) {
@@ -28,8 +30,9 @@ func TestRingConfig_DefaultConfigToLifecyclerConfig(t *testing.T) {
 	expected.FinalSleep = cfg.FinalSleep
 	expected.ReadinessCheckRingHealth = false
 	expected.HeartbeatPeriod = 15 * time.Second
+	expected.RingTokenGenerator = ring.NewRandomTokenGenerator()
 
-	assert.Equal(t, expected, cfg.ToLifecyclerConfig())
+	assert.Equal(t, expected, cfg.ToLifecyclerConfig(log.NewNopLogger()))
 }
 
 func TestRingConfig_CustomConfigToLifecyclerConfig(t *testing.T) {
@@ -44,16 +47,18 @@ func TestRingConfig_CustomConfigToLifecyclerConfig(t *testing.T) {
 	cfg.ExcludedZones = []string{"zone-a", "zone-b"}
 	cfg.TokensFilePath = "/tokens.file"
 	cfg.NumTokens = 1234
-	cfg.InstanceID = "instance"
+	cfg.InstanceID = "instance-10"
 	cfg.InstanceInterfaceNames = []string{"abc"}
 	cfg.InstancePort = 1111
 	cfg.InstanceAddr = "1.2.3.4"
-	cfg.InstanceZone = "zone-X"
+	cfg.InstanceZone = "zone-c"
 	cfg.UnregisterOnShutdown = true
 	cfg.ObservePeriod = 10 * time.Minute
 	cfg.MinReadyDuration = 3 * time.Minute
 	cfg.FinalSleep = 2 * time.Minute
 	cfg.ListenPort = 10
+	cfg.TokenGenerationStrategy = spreadMinimizingTokenGeneration
+	cfg.SpreadMinimizingZones = []string{"zone-a", "zone-b", "zone-c"}
 
 	// The lifecycler config should be generated based upon the ingester ring config
 	expected.RingConfig.KVStore.Store = "memberlist"
@@ -85,5 +90,76 @@ func TestRingConfig_CustomConfigToLifecyclerConfig(t *testing.T) {
 	expected.Addr = cfg.InstanceAddr
 	expected.ListenPort = cfg.ListenPort
 
-	assert.Equal(t, expected, cfg.ToLifecyclerConfig())
+	logger := log.NewNopLogger()
+	tokenGenerator, err := ring.NewSpreadMinimizingTokenGenerator(expected.ID, expected.Zone, cfg.SpreadMinimizingZones, logger)
+	require.NoError(t, err)
+	expected.RingTokenGenerator = tokenGenerator
+
+	assert.Equal(t, expected, cfg.ToLifecyclerConfig(logger))
+}
+
+func TestRingConfig_CustomTokenGenerator(t *testing.T) {
+	const (
+		instanceID   = "instance-10"
+		instanceZone = "zone-a"
+		wrongZone    = "zone-d"
+	)
+	spreadMinimizingZones := []string{"zone-a", "zone-b", "zone-c"}
+
+	tests := map[string]struct {
+		zone                    string
+		tokenGenerationStrategy string
+		spreadMinimizingZones   []string
+		expectedResultStrategy  string
+		expectedError           bool
+	}{
+		"spread-minimizing and correct zones give a SpreadMinimizingTokenGenerator": {
+			zone:                    instanceZone,
+			tokenGenerationStrategy: spreadMinimizingTokenGeneration,
+			spreadMinimizingZones:   spreadMinimizingZones,
+			expectedResultStrategy:  spreadMinimizingTokenGeneration,
+		},
+		"spread-minimizing and a wrong zone give a RandomTokenGenerator": {
+			zone:                    wrongZone,
+			tokenGenerationStrategy: spreadMinimizingTokenGeneration,
+			spreadMinimizingZones:   spreadMinimizingZones,
+			expectedError:           true,
+		},
+		"random gives a RandomTokenGenerator": {
+			zone:                    instanceZone,
+			tokenGenerationStrategy: randomTokenGeneration,
+			expectedResultStrategy:  randomTokenGeneration,
+		},
+		"unknown token generation strategy gives nil": {
+			zone:                    instanceZone,
+			tokenGenerationStrategy: "bla-bla-tokens",
+			expectedError:           true,
+		},
+	}
+
+	for _, testData := range tests {
+		cfg := RingConfig{}
+		cfg.InstanceID = instanceID
+		cfg.InstanceZone = testData.zone
+		cfg.TokenGenerationStrategy = testData.tokenGenerationStrategy
+		cfg.SpreadMinimizingZones = testData.spreadMinimizingZones
+		err := cfg.Validate()
+		if testData.expectedError {
+			require.Error(t, err)
+			continue
+		}
+		require.NoError(t, err)
+		lifecyclerConfig := cfg.ToLifecyclerConfig(log.NewNopLogger())
+		if testData.expectedResultStrategy == randomTokenGeneration {
+			tokenGenerator, ok := lifecyclerConfig.RingTokenGenerator.(*ring.RandomTokenGenerator)
+			assert.True(t, ok)
+			assert.NotNil(t, tokenGenerator)
+		} else if testData.expectedResultStrategy == spreadMinimizingTokenGeneration {
+			tokenGenerator, ok := lifecyclerConfig.RingTokenGenerator.(*ring.SpreadMinimizingTokenGenerator)
+			assert.True(t, ok)
+			assert.NotNil(t, tokenGenerator)
+		} else {
+			assert.Nil(t, lifecyclerConfig.RingTokenGenerator)
+		}
+	}
 }